Linked the Video Queue Thumbnail Images
I still have to figure out how to render the main video image. It seems like it should have been working with props.displayedVideo.image because I was able to console.log that to the right value so I think it could be a component lifecycle issue. Except if it was, would it not have been happening before with the other API? hmmm